+++ src/fastd3/mesh_interpolator_d3.py
@@ -35,7 +35,7 @@ class MeshInterpolatorD3(MeshInterpolator):
 
         rho_flat_accum = torch.zeros(
             (nx * ny * nz, n_channels), 
-            dtype=torch.float32, 
+            dtype=torch.float64, 
             device=self._device
         )
